Skripte/rag-search/requirements.txt
# LlamaIndex Core
llama-index-core>=0.11.0

# Vector Store
llama-index-vector-stores-qdrant>=0.3.0
qdrant-client>=1.7.0
fastembed>=0.4.0

# Embeddings & LLM
llama-index-embeddings-openai>=0.2.0
llama-index-llms-openai>=0.3.0
openai>=1.0.0

# Reranking
llama-index-postprocessor-cohere-rerank>=0.5.0

# Document Parsing
llama-index-readers-file>=0.2.0
llama-parse>=0.5.0

# Utilities
python-dotenv>=1.0.0
